Keep the style catalog apart from the policy that reads it
Colours, marker sizes and font names change for reasons that have nothing to
do with what a batch holds, and they sat in the same file as the rules that
decide it. The catalog and the style identities move to a module of their own.

The projection test moves out of the traffic tests for the same reason: it
asks what the display draws over time, not what one delta becomes.
