+++ crates/pilotage-presentation/src/tests.rs
@@ -3,6 +3,7 @@
 mod advisory;
 mod layer;
 mod ownship;
+mod projection;
 mod traffic;
 mod vertical;
 mod weather;
