+++ crates/pilotage-presentation/src/weather.rs
@@ -6,7 +6,7 @@ use airmass_core::{FlightCategory, WeatherReportType};
 use airmass_geojson::{FeatureDelta, TextReportFeature, WeatherFeatureId, WeatherFeatureKey};
 
 use crate::layer::WEATHER_REPORT_LAYER_ID;
-use crate::policy::{
+use crate::style::{
     WEATHER_IFR_STYLE, WEATHER_LIFR_STYLE, WEATHER_MVFR_STYLE, WEATHER_UNKNOWN_STYLE,
     WEATHER_VFR_STYLE,
 };
